+++ sdk/storage/azdatalake/internal/path/models.go
@@ -19,6 +19,7 @@ import (
 type DeleteOptions struct {
 	// AccessConditions contains parameters for accessing the path.
 	AccessConditions *AccessConditions
+	Paginated        *bool
 }
 
 func FormatDeleteOptions(o *DeleteOptions, recursive bool) (*generated.LeaseAccessConditions, *generated.ModifiedAccessConditions, *generated.PathClientDeleteOptions) {
@@ -27,6 +28,9 @@ func FormatDeleteOptions(o *DeleteOptions, recursive bool) (*generated.LeaseAcce
 	}
 	if o == nil {
 		return nil, nil, deleteOpts
+	}
+	if o.Paginated != nil {
+		deleteOpts.Paginated = o.Paginated
 	}
 	leaseAccessConditions, modifiedAccessConditions := exported.FormatPathAccessConditions(o.AccessConditions)
 	return leaseAccessConditions, modifiedAccessConditions, deleteOpts