+++ sdk/storage/azdatalake/internal/testcommon/clients_auth.go
@@ -36,6 +36,7 @@ const (
 	DataLakeEncryptionScopeEnvVar = "DATALAKE_AZURE_STORAGE_ENCRYPTION_SCOPE"
 	SubscriptionID                = "SUBSCRIPTION_ID"
 	ResourceGroupName             = "RESOURCE_GROUP_NAME"
+	ObjectIdEnvVar                = "AZURE_APP_OBJECT_ID"
 )
 
 const (